added walkthrough transform and renamed
