+++ src/redux/actions/walkthroughActions.ts
@@ -1,5 +1,5 @@
 import { REGISTER_TOOLTIP } from '../constants/constants';
-import { Walkthrough } from '../reducers/tooltipsReducer';
+import { Walkthrough } from '../reducers/walkthroughReducer';
 
 export const registerTooltip = (walkthrough: Walkthrough) => ({
   payload: walkthrough,
