+++ src/redux/reducers/index.js
@@ -9,7 +9,7 @@ import user from './userReducer';
 import customTabsReducer from './customTabsReducer';
 import editorReducer from './editorReducer';
 import cacheReducer from './cacheReducer';
-import tooltipsReducer from './tooltipsReducer';
+import walkthroughReducer from './walkthroughReducer';
 
 export default combineReducers({
   account: accountReducer,
@@ -22,5 +22,5 @@ export default combineReducers({
   communities,
   user,
   cache: cacheReducer,
-  tooltips: tooltipsReducer,
+  walkthrough: walkthroughReducer,
 });
